[Refactor][Tests] trim test_reduction_alignment.py to direct LogSoftmax FLOP regression
Removes the manifest-string-eval and metadata-pinning tests introduced in
PR #1235 and reduces the alignment file to the minimum that verifies the
only behavior change PR #1235 actually shipped: the LogSoftmaxFwdOp
eval_roofline() FLOP correction (5 * M * N).

- Delete tests/ops/test_reduction_alignment.py and its `_eval_roofline_expr`
  helper (eval()-on-manifest-strings security smell).
- Add a direct regression test `test_log_softmax_eval_roofline_flops_5mn`
  to tests/ops/test_softmax.py — constructs LogSoftmaxFwdOp, runs forward,
  asserts flops == 5 * M * N with no manifest indirection.
- Move 3 behavior tests verbatim into tests/ops/test_logical_reduce.py:
  test_logical_reduce_accepts_bool, test_count_nonzero_returns_int64,
  test_logical_reduce_returns_bool.
- Drop test_reduction_constructs_for_manifest_dtypes (every spec-only op
  is already covered for fp16/bf16/fp32 in test_reduce.py and for bool /
  int64 in test_logical_reduce.py).
- Drop test_reduction_signature_matches_manifest (signature parity is the
  validator's job, not a per-family pytest).

Diff confined to tests/ops/. tileops/manifest/reduction.yaml and
tileops/ops/reduction/_softmax_base.py are byte-identical.
